Implemented Oct Tree Traversal and Rendering Methods - Don't Work Yet
Bounding Objects within the Octants don't draw as intended. I believe
this has something to do with how managers handle render rather than
individual rendering them selves. Also It can create a root node and the
second level of children but it won't send the data to the children and
delete it from the parent.