Added a Read Me document and Deleted Old Version
I created a read me that explains the projects purpose its controls and
how we applied the topics. I also removed the old version of the project
from the repo since all of our work is now combined. This will also
solve any confusion if the professor or grader is looking for our latest
working version